# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tolumo API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /enrollments:
    get:
      operationId: listEnrollments
      tags: [enrollments]
      summary: List current student's enrollments with progress
      responses:
        "200":
          description: Enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
    post:
      operationId: createEnrollment
      tags: [enrollments]
      summary: Enroll current student in a module
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollmentInput"
      responses:
        "201":
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings (role-scoped; students see theirs, tutors see theirs, admin sees all)
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Book a one-on-one tutorial session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update a booking's status or notes
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /acquisitions:
    get:
      operationId: listAcquisitions
      tags: [agents]
      summary: List acquisitions (sub-agent sees own; super-agent/admin see all)
      responses:
        "200":
          description: Acquisitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Acquisition"
    post:
      operationId: createAcquisition
      tags: [agents]
      summary: Record a new student acquisition/lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcquisitionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Acquisition"
  /acquisitions/{id}:
    patch:
      operationId: updateAcquisition
      tags: [agents]
      summary: Update an acquisition's status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcquisitionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Acquisition"
...
components:
  schemas:
...
    Enrollment:
      type: object
      required: [id, moduleId, moduleTitle, moduleCode, year, enrolledAt, lessonCount, completedLessons]
      properties:
        id: { type: integer }
        moduleId: { type: integer }
        moduleTitle: { type: string }
        moduleCode: { type: string }
        year: { type: integer }
        enrolledAt: { type: string }
        lessonCount: { type: integer }
        completedLessons: { type: integer }
        imageUrl: { type: ["string", "null"] }
    EnrollmentInput:
      type: object
      required: [moduleId]
      properties:
        moduleId: { type: integer }
    Booking:
      type: object
      required: [id, studentId, tutorId, studentName, tutorName, scheduledAt, status]
      properties:
        id: { type: integer }
        studentId: { type: integer }
        tutorId: { type: integer }
        studentName: { type: string }
        tutorName: { type: string }
        moduleId: { type: ["integer", "null"] }
        moduleTitle: { type: ["string", "null"] }
        scheduledAt: { type: string }
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        notes: { type: ["string", "null"] }
    BookingInput:
      type: object
      required: [tutorId, scheduledAt]
      properties:
        tutorId: { type: integer }
        moduleId: { type: integer }
        scheduledAt: { type: string }
        notes: { type: string }
    BookingUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        scheduledAt: { type: string }
        notes: { type: string }
    Acquisition:
      type: object
      required: [id, subAgentId, subAgentName, studentName, status, commissionAmount, createdAt]
      properties:
        id: { type: integer }
        subAgentId: { type: integer }
        subAgentName: { type: string }
        studentName: { type: string }
        studentEmail: { type: ["string", "null"] }
        status:
          type: string
          enum: [lead, contacted, enrolled, paid]
        commissionAmount: { type: number }
        createdAt: { type: string }
    AcquisitionInput:
      type: object
      required: [studentName]
      properties:
        studentName: { type: string, minLength: 1 }
        studentEmail: { type: string }
        status:
          type: string
          enum: [lead, contacted, enrolled, paid]
        commissionAmount: { type: number }
    AcquisitionUpdate:
      type: object
      properties:
        studentName: { type: string, minLength: 1 }
        studentEmail: { type: string }
        status:
          type: string
          enum: [lead, contacted, enrolled, paid]
        commissionAmount: { type: number }
...