# lib/api-zod/src/generated/types/acquisitionUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Tolumo API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AcquisitionUpdateStatus } from './acquisitionUpdateStatus';

export interface AcquisitionUpdate {
  /** @minLength 1 */
  studentName?: string;
  studentEmail?: string;
  status?: AcquisitionUpdateStatus;
  commissionAmount?: number;
}