# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Tolumo API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  studentId: number;
  tutorId: number;
  studentName: string;
  tutorName: string;
  /** @nullable */
  moduleId?: number | null;
  /** @nullable */
  moduleTitle?: string | null;
  scheduledAt: string;
  status: BookingStatus;
  /** @nullable */
  notes?: string | null;
}