# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Tolumo API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  tutorId: number;
  moduleId?: number;
  scheduledAt: string;
  notes?: string;
}